[EICNET-1545] chore: Handle subscriptions with the message bus
