+++ lib/modules/eic_messages/eic_messages.services.yml
@@ -1,7 +1,10 @@
 services:
+  eic_messages.queued_message_checker:
+    class: Drupal\eic_messages\Util\QueuedMessageChecker
+    arguments: [ '@datetime.time', '@config.factory', '@current_user', '@entity_type.manager' ]
   eic_messages.message_bus:
     class: Drupal\eic_messages\Service\QueueMessageBus
-    arguments: [ '@datetime.time', '@config.factory', '@current_user', '@entity_type.manager' ]
+    arguments: [ '@eic_messages.queued_message_checker' ]
     tags:
       - { name: service_collector, tag: message_producer, call: addHandler }
