+++ lib/modules/eic_messages/modules/eic_message_subscriptions/eic_message_subscriptions.services.yml
@@ -1,15 +1,19 @@
 services:
   eic_message_subscriptions.event_subscriber:
     class: Drupal\eic_message_subscriptions\EventSubscriber\MessageSubscriptionEventSubscriber
-    arguments: ['@eic_messages.message_creator.comment', '@eic_messages.message_creator.group_content', '@eic_messages.message_creator.node', '@message_subscribe.subscribers']
+    arguments: [ '@eic_message_subscriptions.message_creator', '@message_subscribe.subscribers', '@eic_messages.queued_message_checker' ]
     tags:
       - { name: event_subscriber }
 
   eic_message_subscriptions.helper:
     class: Drupal\eic_message_subscriptions\MessageSubscriptionHelper
 
+  eic_message_subscriptions.message_creator:
+    class: Drupal\eic_message_subscriptions\Service\SubscriptionMessageCreator
+    arguments: [ '@current_user' ]
+
   eic_message_subscriptions.flag_event_subscriber:
     class: Drupal\eic_message_subscriptions\EventSubscriber\FlagEventSubscriber
-    arguments: ['@event_dispatcher']
+    arguments: [ '@event_dispatcher' ]
     tags:
       - { name: event_subscriber }
