+++ lib/modules/eic_messages/src/Service/MessageBusInterface.php
@@ -2,23 +2,10 @@
 
 namespace Drupal\eic_messages\Service;
 
-use Drupal\message\MessageInterface;
-
 /**
  * Interface MessageCreatorInterface to implement in MessageCreator services.
  */
 interface MessageBusInterface {
-
-  /**
-   * Checks if the given message should be created or not.
-   *
-   * @param \Drupal\message\MessageInterface $message
-   *   The message to be created.
-   *
-   * @return bool
-   *   Whether to create the message or not.
-   */
-  public function shouldCreateNewMessage(MessageInterface $message): bool;
 
   /**
    * @param \Drupal\message\Entity\Message|array $message
