+++ lib/modules/eic_messages/src/Util/ActivityStreamMessageTemplates.php
@@ -6,6 +6,7 @@ use Drupal\Core\Entity\ContentEntityInterface;
 use Drupal\eic_messages\MessageIdentifierInterface;
 use Drupal\eic_messages\MessageTemplateTypes;
 use Drupal\message\MessageTemplateInterface;
+use InvalidArgumentException;
 
 /**
  * Helper class for activity stream message templates.
@@ -79,23 +80,6 @@ final class ActivityStreamMessageTemplates implements MessageIdentifierInterface
   }
 
   /**
-   * Returns an activity stream message template for a given entity.
-   *
-   * @param \Drupal\Core\Entity\ContentEntityInterface $entity
-   *   The entity object.
-   *
-   * @return string
-   *   The template name.
-   */
-  public static function getTemplate(ContentEntityInterface $entity): string {
-    if (!isset(self::$templates[$entity->getEntityTypeId()][$entity->bundle()])) {
-      throw new \InvalidArgumentException('Invalid entity / bundle provided');
-    }
-
-    return self::$templates[$entity->getEntityTypeId()][$entity->bundle()];
-  }
-
-  /**
    * Checks if an entity has activity message template.
    *
    * @param \Drupal\Core\Entity\ContentEntityInterface $entity
@@ -108,7 +92,7 @@ final class ActivityStreamMessageTemplates implements MessageIdentifierInterface
     try {
       self::getTemplate($entity);
     }
-    catch (\InvalidArgumentException $e) {
+    catch (InvalidArgumentException $e) {
       return FALSE;
     }
 
@@ -116,11 +100,26 @@ final class ActivityStreamMessageTemplates implements MessageIdentifierInterface
   }
 
   /**
+   * Returns an activity stream message template for a given entity.
+   *
+   * @param \Drupal\Core\Entity\ContentEntityInterface $entity
+   *   The entity object.
+   *
+   * @return string
+   *   The template name.
+   */
+  public static function getTemplate(ContentEntityInterface $entity): string {
+    if (!isset(self::$templates[$entity->getEntityTypeId()][$entity->bundle()])) {
+      throw new InvalidArgumentException('Invalid entity / bundle provided');
+    }
+
+    return self::$templates[$entity->getEntityTypeId()][$entity->bundle()];
+  }
+
+  /**
    * {@inheritdoc}
    */
   public static function getMessageTemplatePrimaryKeys(MessageTemplateInterface $message_template) {
-    $primary_keys = [];
-
     // Get the message template type.
     $message_template_type = $message_template->getThirdPartySetting('eic_messages', 'message_template_type');
 
@@ -130,11 +129,9 @@ final class ActivityStreamMessageTemplates implements MessageIdentifierInterface
 
     // We assume all stream messages reference a node.
     // @todo Define if all stream templates should also include executing user.
-    $primary_keys = [
+    return [
       'field_referenced_node',
     ];
-
-    return $primary_keys;
   }
 
 }
